Experimenting with the API output
These will change.I will ask for more feedback once I have incorporated more data. Though, people are always welcome to chime in at any stage.

I am not sold on the elections being a dictionary, it might be an intermediate step before it is a list.

This is not the most efficient, but gets us started.

I added some in-line documentation if people are interested.