Added bootstrapping scripts and updated README
The getting started docs were miserable out of date, and there is no
longer anything that needs to be kept secret for running a local copy.
This change adds a bootstrap script that, provided all requirements
listed in the README are installed, will clone the repos, set up env
variables, and add a tmuxinator config to start both apps together and
in the background. It also adds a Vagrant box and provisioning script
that the bootstrap can be run on.