Update first and last filing dates for committees.
Note: Based on the recommendation of FEC, this patch only considers
receipt dates from F1 filings. Since we're doing an `INNER JOIN` to combine
`dimcmte` and `dimcmteproperties`, committees that have never filed an F1
are excluded. We could switch to a `LEFT JOIN`, but then we'd have a bunch
of committees with `NULL` dates and values from `dimcmteproperties`.

Thoughts @lindsayyoung @arowla?

[Resolves #624]
[Resolves #649]