Harmonize committee names across views.
We currently build committee names different in the committee and
committee detail views versus the full-text search view. This means that
it's possible to find committee names in the typeahead search, click on
a name, and not find that name in the committee results page. This patch
updates the full-text search view to use the same logic to build
committee names as used in the other committee views. This also replaces
some indirection using `load_order` with `select distinct` plus `order
by`, which is more consistent with the rest of the import scripts.

[Resolves #676]