Move form filter to correct join.
Based on correspondence with FEC, we should only use F1 filings to
determine original filing dates, but we can use the most recent filings
regardless of form type for other committee properties. I got this
almost right in #678 but applied the F1 filter to the most recent
committee properties rather than the original properties. This patch
moves the filter to the right join.

Thanks @lindsayyoung and @arowla for identifying the issue.