Restore correct `election_years` and `active_through` columns.
When we fixed candidate party descriptions by altering the join between
`dimcand` and `dimcandoffice`, we also introduced a regression, such
that each candidate had at most one year in `election_years`. To select
both the correct row in `dimcandoffice` to determine party affiliation
and the correct rows to determine election years, we need two joins.
This patch adds a second join, and harmonizes the candidate views to use
the same logic.