Remove duplicated reports.
Previously, we were joining `dimcmte` on `dimcmtetpdsgn` to generate
views for committee reports. Because committees may have multiple
designations, this was in some cases creating duplicate reports. Since
we only used the join on `dimcmtetpdsgn` to fetch the committee type,
and since this value was neither imported into SQLAlchemy nor serialized
to the API, the simplest way to handle the duplication was to remove the
join.