Fix sorting on candidate search view.
When using `subqueryload` for eager loading with the `CandidateSearch`
resource, we have to specify which model we're sorting on, e.g.
`Candidate.candidate_id` instead of "candidate_id". This patch adds an
optional `model` argument to the `fetch_page` helper, which retrieves
the column property from the model if provided. We also now throw a 400
if a user tries to sort on a column that doesn't belong to the relevant
model.

h/t @msecret