Remove name fulltext view.
The name fulltext view is more or less the candidate and committee
fulltext views concatenated together. We don't need to build these rows
twice; the separate candidate and committee fulltext views are adequate
as they are. This patch deletes the combined view and the logic used to
handle the complexities of using a single view to store two different
kinds of rows.