Use consistent endpoints for candidate and committee history resources.
For the webapp (and potentially other consumers of the API), we need to
be able to look up information about nested resources at a particular
time: e.g., information about all of a candidate's associated committees
in the 2011-2012 election cycle. We currently support these nested
history lookups for committee history, but not for candidate history,
exposing an edge case on the webapp where we can occasionally show the
wrong information for a candidate on her/his committee page. This patch
adds consistent endpoints for nested history resources.