Add queued incremental aggregates.
Aggregates over Schedule A and B receipts can be expensive to build from
scratch. This patch implements queued incremental aggregates, such that
updates to the `sched_a` table are pushed into a queue, and aggregates
are incrementally updated on demand:

* Create initial aggregates (slow!)
* Store all changes to `sched_a` in queue tables
* As needed, update all aggregates and clear queues

This approach adds the overhead of copying added and deleted rows to the
queue tables, but this is much less expensive than updating all
aggregates in real time or rebuilding aggregates from scratch.

Many thanks to @arowla for suggesting this approach.