Use Flask-CORS for more robust CORS configuration.
The API currently uses the `Access-Control-Allow-Origin` header, but not
the `Access-Control-Allow-Headers` header. This leads to CORS errors
when using Safari, which sends the `Accept-Encoding` header with its
`OPTIONS` requests. Rather than allowing the `Accept-Encoding` header,
this patch switches to Flask-CORS for CORS configuration, which has
better defaults and is easier to configure than settings CORS headers
manually.

[Resolves https://github.com/18F/openFEC-web-app/issues/462]