Allow filtering receipts on multiple committees.
When filtering on multiple committees, the query planner can't take
advantage of the compound indexes on `sched_a_sk` and other receipts
columns, resulting in very slow queries. This patch creates a separate
subquery for each committee and combines the results with `UNION ALL`.
Note that both the subqueries and the combined query must apply the same
filters and `ORDER BY` rules.

h/t @arowla for suggesting subqueries