Show all candidates in election views.
Currently, we use an `INNER JOIN` to join candidates on totals for the
election summary and detail views. This has the unwanted effect of
excluding candidates with no totals from these views. This patch uses
a `LEFT OUTER JOIN` and updates the queries accordingly, such that all
relevant candidates are included.

Related to #646