Restrict Schedule E aggregates to form F3X.
This is a temporary fix for #1226. As described in the issue, Schedule E
transactions from 24- and 48-hour reports aren't currently grouped by
unique transaction, such that multiple amended versions of the same
transaction can exist in the database. In the long term, @paulclark2 and
@jwchumley will work on aggregating Schedule E transactions across
amendments, but as a short-term fix, this patch simply restricts
Schedule E aggregates to transactions filed on form F3X.