Temporarily fetch committee type from detail table.
Due to a mismatch in active cycles between `dimcmteproperties` and
`vw_filing_history`, some committee reports and totals return 404s on
attempting to resolve the committee type when a cycle is specified. This
patch temporarily falls back to resolving committee type from the detail
table when no matching results can be found for the specified cycle.

Note: this should be reverted after the underlying issue described in
https://github.com/18F/openFEC-web-app/issues/875 is resolved.