Use committee designation on linkage table.
Join candidates to their principal committees using committee
designation values from the linkage table. This ensures that we're
fetching the most recent principal committee, not the most recent
committee that was ever a principal committee. This also fixes a few
candidates for whom principal committee designations are correct in the
linkage table but missing in the committee designation table.

Related to #1299.