Filter on derived `is_individual` column.
Now that we're deriving the `is_individual` column on the Schedule A
table on load, we can simplify our in-app filtering rules. This should
also give us faster queries when filtering on `is_individual`, since
we'll be filtering on a single indexed column instead of several
columns, some of which aren't indexed.

Noticed while debugging #1330.