Censor duplicate committee IDs in itemized tables.
As described in #1306 and #1330, committees sometimes enter the same values for
`contbr_id` and `cmte_id` (for receipts) and for `recipient_cmte_id` and
`cmte_id` (for disbursements). In these cases, values of `contbr_id` and
`recipient_cmte_id` should be ignored for receipts and disbursements,
respectively. This patch creates `clean_contbr_id` and
`clean_recipient_cmte_id` columns, which are displayed in the API
instead of the original `contbr_id` and `recipient_cmte_id` columns.

[Resolves #1306]
[Resolves #1330]