Ensure itemized queues are unique on corresponding primary key.
Nightly updates on itemized tables fail when the same item has been
added or modified more than once within a single update cycle. To fix,
this patch adds primary key constraints on the queue tables and deletes
any existing records in the queues by primary key on insert.

Note: We'll be able to express this more concisely as an upsert once
postgres 9.5 is available on RDS.

[Resolves #1392]