Ensure stable sort on aggregate tables.
* Disable multi-column sort
* Always sort on unique keys on aggregate tables

By sorting aggregate tables on a unique key as well as the
user-specified sort key, results are returned in a stable order even in
the presence of duplicates on the selected sort column. Note: this
requires that tables have composite indexes on each possible sort column
and the unique column.

[Resolves #1361]