Fix and clean up statutory candidate flag.
As implemented, the statutory candidate flag includes a number of false
negatives due to a `group by` on `cand_status`, combined with a
`distinct` on `cand_id`. The `group by` yields potentially many rows per
candididate, with some `true` and some `false`; the `distinct` selects
just one of these rows, which could be a `false` when other rows are
`true`. This patch drops these two expressions. And now that the
grouping logic is the same as in a previous select, it also combines
selects to make the view slightly faster to build.

* Drop group by on `cand_status`
* Drop unnecesary distinct on `cand_id`
* Move status aggregate into existing select
* Simplify case to plain &&