Add manifest files for cf-ssh access
This changeset adds additional manifest files so that we are able to use the `cf-ssh` cli tool in order to connect directly to our servers.

NOTE:  This is just a stop-gap measure fow now, we will want to improve this in the near future, but we need to be able to access these machines now.  The current problem appears to be that you are unable to use a manifest that inherits from another one when using the `cf-ssh` cli-tool, so that will be a place to start investigating further.