Updates to our contributing guidelines
This changeset updates our CONTRIBUTING.md documentation with
up-to-date information around our team practices based on recent
conversations we have had internally.

There are now two formal teams at 18F working with the FEC on related
projects and we are largely working within the same repositories for
project-related work. Based on recent conversations between the two
teams, we are committed to supporting one another to achieve success
and produce the highest quality work we can.

These changes reflect the agreements and commitments we
have made to one another. As our needs and/or the needs of the
project change, we are also committed to updating these team
practices as necessary to keep them current.