Updated subset.dump file
This changeset updates our test subset of data with a sampling of data taken around midnight of May 7, 2016. This subset was pulled from the `dev` database.

There have been numerous updates to the data set since the last subset was pulled so we felt it was time to refresh it.