Rename transaction_year field
Part of #1735

This changeset renames the field we originally went with for partitioning the itemized schedule A and B tables. We had received feedback that `transaction_year` was a bit confusing and have decided to go with `two_year_transaction_period` instead to be more descriptive of what this value represents.