Account for CF target in deploy task
Addresses #1738

This changeset ensures that we only authenticate with cloud.gov if the login environment variables are set (which should only be in the case of a bot, e.g., Travis CI), but that we always change our target organization and space to match what was detected and/or passed in.

h/t to @adborden on accounting for the targeting step!