Fix Schedule A and B updates
This changeset addresses an issue with failing updates for the Schedule A and B tables.  The update process did not take into account the table partitioning and was still basing the checks off of the report year (`rpt_yr`) instead of the new calculated value (`two_year_transaction_period`).  In order to fix this, I had to do the following:

* Add the `two_year_transaction_period` column to both sets of queue tables for Schedule A and B.
* Remove the `two_year_transaction_period` column from the column factory in the select statement generation for processing updates.

I do not believe this is entirely straightforward so there is a comment in the code that refreshes the child tables during an update process to help explain the reasoning behind what is happening.

h/t to @LindsayYoung for working through the approach with me!