Adjusts aggregate table processing to make use of temporary tables
Fixes #1775

This changeset adjusts the processing for the aggregate tables to make use of temporary tables first. Prior to this, any updates to the aggregates in the database would cause the site to not yield the data that we expect to see.  We would like to achieve as little downtime as possible and the rest of our database processing makes use of creating temporary tables and views to accomplish this, so we are making use of them here as well.