Add support for renaming indexes
When partitioning the itemized Schedule A and B tables, the indexes of the existing tables were being dropped and recreated, but this impacted the existing tables before they were replaced.  As partitioning these tables takes a long time with the full FEC data set, this has huge impact on database performance if anyone or anything is accessing the data in these tables.

This changeset alleviates this by renaming the indexes once the tables themselves are renamed so that they are not removed too early.