Added a step to the transaction tests to ensure the queues are clear
NOTE:  This is definitely not an ideal state, but it seems that some of our tests do have interdependencies so we need this workaround for the time being.  In the future we should ensure we have proper setup and teardown methods so that each test truly runs in isolation of all the others with no unintended side effects.