Address issue with celerybeat-schedule
This changeset updates Celery to the latest patch level version and addresses a small problem we ran into during recent deploys.  If you have a local copy of the `celerybeat-schedule` file, it will go up with the deploy to cloud.gov and cause the `celery-beat` app to error out upon starting.  Adding this file to the `.gitignore` and `.cfignore` files will prevent this from happening.