Modify mirror support to redirect all read traffic
This changeset modifies the mirror support currently in place to redirect all of our read traffic to any configured follower databases. The code checking for specific tasks might still be useful in the future, so I added a little extra support to toggle restricting traffic to tasks only that can be adjusted via an environment variable.

This also updates SQLAlchemy to the latest patch level release of the 1.0 series.