Adds committee names back into itemized schedule downloads
This changeset addresses #1940 by adding support for related fields back into our exports on the itemized tables.  It does this by defining an extra attribute in our schema Meta objects that is ignored by everything but our export code.  This extra attribute allows you to set additional columns that are found in nested models related to the parent model you are trying to export.  If any are defined, the export functionality will follow the relationship to pull the information out and display it in the output with the specified label.  Documentation in the schema provides more information on how to do this.

This changeset also modifies our Schedule E model by removing the `committee_name` field set directly to the `cmte_nm` field in the table.  This field can sometimes be `null`, so we should be pulling it from the related `CommitteeHistory` model instead, just like the Schedule A and B models already do.