Reverted sorting to maintain original order and added support to insert columns in specific spots
The sorting we had initially for the columns in the export output is a desired order, but the committee names would make sense to keep closer to the front. This change allows for that and does so in a way that we can leverage it for any future additions as well for columns that come from related models.