Add committee names to itemized aggregate exports
This changeset adds support for displaying committee names in the CSV exports of the itemized aggregate tables.  There is a small amount of technical debt taken on here as I had to incorporate the relationships across multiple models and the existing schema creation convenience methods do not account for this easily.  There might be a way to simplify this in the future, but I also did not want to impact the existing schema building that we have.

This may also expose some additional committee information in the API endpoints, but we can exclude those if need be.