Adds secondary sorting for itemized resources in the API
This changeset adds support for adding a secondary sort column for all itemized resources to help address the issue with paging through records that have `null` values for the initial sort column (e.g., `expenditure_date` for `ScheduleE`).  This does not fully fix the issue though as the instructions and usage still indicate that if you want to preserve ordering in your result you need to supply the `last_x` column (with `x` being the column you are sorting on) in addition to the `last_index` parameter.  This is still being hammered out and discussed in issue #1960.