Added sort_reverse_nulls flag, updated documenation, and added tests
This changeset introduces a new flag called `sort_reverse_nulls` that can be used to reverse the order in which nulls are presented in a result listing.  By default, Postgres will show records with null values on a sorted column last when sorting in ascending order or first when sorting by descending order.  This flag lets us reverse this behavior if desired.