Remove unused code
remap_archived_murs_citations was added in order to fix archived murs
that had already been loaded into elasticsearch. With the decision to
release only current MURs (and consequently to remove archived MURs from
elasticsearch), it is no longer needed. If and when we decide to release
archived MURs, we will be running load_archived_murs again, which will
has the citations correctly mapped.