Adds support for generating new itemized child tables
This changeset adds a management command to support creating new child tables to an existing partition of Schedule A and Schedule B data. It will check first before just creating the new table so existing data is not wiped out.

This also includes the beginning of a test suite for the partitioning code.