Specify schema for MURs explicitly
This is a recommended practice for production implementations. It allows
us to document the schema better.

We will eventually do the same for the other document types and remove the
_default_ mappings.

Turn off indexing for fields like document_ids, URLs, etc as they tend
to result in false positives for searches.