Adds support to sort by date for entity charts
This changeset adds support for sorting by date in our entity charts.  It accomplishes this by deriving a date from the `month` and `year` columns that are created in the `ofec_entity_chart_mv` materialized view (and always assuming the first of the month as the day).  Note that we already have a `date` property on the Python side in the `EntityReceiptDisbursementTotals` model, so I did not want to disrupt anything currently using that.  Hence the name for this new column and field being `chart_date`.

h/t to @LindsayYoung for the implementation idea!