Updated the schedule A and B triggers to leverage the views
This modification allows us to leverage the structure of the data in the views, which is what we have come to rely upon an expect, but make use of triggers when the source data is inserted/updated/removed.  In order to accomplish this I had to create before and after triggers depending on what the operation was, and in the case of updates, use both.

This in turn has required some additional changes to ensure that a contribution or disbursement date is present on the records, or these operations will fail, so I added those properties to the test models and factories.