Add support and instructions for cf ssh
This changeset adds support and instructions for running `cf ssh` in the GovCloud environment.  While it is easy and straightforward to SSH into a running application, there is a little bit of environment setup to do once you are in.  This script makes it easy to take care of these things with one command.