Update some of the backend tooling for the API
This changeset updates some of the tooling and infrastructure in the web app to bring things up to date. It includes the following:

* Cleanup of .travis.yml
* Updates several Python dependencies
* Removal of all references to Sentry/Raven
* Updates cloud.gov Python runtime to 3.5.3
* Updates New Relic environment name in the manifest_feature.yml config
* A handful of documentation updates

The Travis cleanup includes updating nvm to make use of the latest stable release of Node.js, which I have been running locally for a while now with no issues.