Fix refresh_child method to account for column type mis-match
This changeset addresses an issue with the refresh_child method not accounting for the `schedule_type` column being set to `unknown` in the schedule A and B queue tables.  The previous work done to fix this issue did not take into account the nightly refresh component of the partition code, only the creation of the partition.

Additional notes have been added to some of the comments as well to explain the purpose and usage of some methods as well.