Improve logging and split out calls for nightly updates
This changeset adds a bit of extra support for logging when the nightly updates are run for the itemized schedule data.  It also splits the methods out a bit more from the one management command into one for each, just like the we did for the partitioning code.  This should provide us with more flexibility as well as more visibility into any problems that are currently being swallowed by the success messages.