Fix itemized schedule trigger errors
This changeset addresses an issue we are currently having with our itemized schedule triggers.  The triggers are setup to select rows from views currently, but they often run into a situation where the selection statement actually returns an empty resultset.  This in turn causes what amounts to an empty record (excluding what we ourselves add) being added to the itemized schedule queues to be processed later, only to fail because of trying to insert records that do not comply with table constraints (e.g., inserting a NULL value into a primary key column).

This puts us in a state of having to perpetually refresh data manually, which requires repartitioning the itemized schedule tables.  There is more work to be done to fix the root cause(s) of this problem, but this will at least prevent errors in our nightly refresh process for the time being.

I have also fixed a couple of issues with the integration tests in this.