Fix nightly update trigger improvements
This changeset addresses a couple of issues we uncovered with our last nightly update fixes.  First, the record was not being returned properly in the trigger function due to incorrect shuffling around of return statements.  Second, we now also track which sub_ids are not found and place them in a secondary queue.  This queue is processed separately and mirrors the functionality of the main queues with the exception that sub_ids will remain in here until the record can be processed successfully from the itemized views we have to process from.

These improvements are a stop gap measure to keep our nightly updates running smoothly until we can improve the overall processing of the itemized schedule data.