Remove RAISE statements from triggers
This changeset removes the `RAISE` statements in the itemized schedule triggers as they appear to be causing GoldenGate to malfunction when they are called.  They are only supposed to write statements to the PostgreSQL log, not throw exceptions, but it seems that GoldenGate is treating them as exceptions anyway and halting processing.