Refactor itemized schedule A/B processing to use "upsert"
This changeset introduces a major refactoring in how we keep track of and process the itemized schedule data.  Instead of using the queue tables like before, we are now leveraging the new `INSERT INTO ... ON CONFLICT ...` functionality that PostgreSQL provides as of 9.5+.  This has allowed us to greatly simplify the triggers we have associated with the itemized schedule A and B data and remove a lot of complicated data processing and management code with the table partitioning we have setup for the API.  It also provides us with updates to the table as the changes happen instead of during a nightly process.

These changes are only for schedule A and B right now, but at least schedule E could also benefit from a similar refactor in the future.