Adding in fix for #2426
Issue #2426 uncovered a need for us to make sure that the materialized views are updated in the correct order, and the refresh script we have does not do that.  This is a semi-temporary fix until we can work in a better solution, but at least it guarentees the materialized views are refreshed properly.

h/t to @jontours, @LindsayYoung, and @anthonygarvan for helping out with this!