Attempting to get inserts and updates into child tables working correctly
This also adds the beginnings of the partition code on the SQL side with a trigger for the master tables.  This trigger has to be manually applied for the time being until we get more worked in, but should eventually be managed by another method in `manage.py` like our other SQL scripts to get run at the appropriate time during a database update/refresh.