Change Schedule C to a materialized view
This changeset readjusts our schedule C data to be available via a materialized view as opposed to accessing the fec_vsum_sched_c_vw view directly.  As we cannot setup triggers to watch the view, this is an easier way to keep the information updated.

This also fixes an issue with the site currently where loan information is not visible because of a reference to the pg_date column, which was not in the original view.