Simplifies schedule A and B triggers and setup to use just rpt_yr
This commit is relatively significant as it also eliminates the get_transaction_year SQL functions.  With the two_year_transaction_period being calculated solely from rpt_yr, which is now coming from a single source of truth (disclosure.f_item_receipt_or_exp), we are able to implement a recommendation from the FEC DB team and simplify this logic.