Remove two year cycle restriction on itemized schedule API endpoints
This changeset removes the two year cycle restriction from the itemized schedule A and B API endpoints to enable searching and querying across multiple cycles.  Very early performance tests indicate that as long as the indexes are setup properly this is reasonably performant, even with large cross sections of data.

**NOTE**
* We still need to provide some type of restriction; simply asking for the whole data set results in a long-running query
* It might be possible to setup a restriction just based on min/max dates alone; that is worth exploring
* Alternatively, we could either allow multiple values for `two_year_transaction_period` or change that parameter to work on groupings of 4, 6, or 8 cycles

Again, this code only drops the restriction in the API for the moment and is only meant for testing and further development.  Work should be actively coordinated with the front end development in the web app to ensure consistency and proper UI/UX support.