Force using Trusty and PostgreSQL 9.6 in Travis CI
We are running into some issues with our tests failing in Travis and currently all signs point to something breaking down in the Travis build.  This changeset forces the build environment to use Trusty (which Travis is in the process of rolling out to all builds regardless), which in turn also supports using PostgreSQL 9.6 (which this project is running on).  We will continue making adjustments as necessary to see what is causing the tests to fail and what might be needed to get them to work.