Send downloads directly to S3
This avoids using the local filesystem, which is good because we have
occasionally run into "No space left on device" errors.

We also avoid creating a zip file and a manifest and instead upload a
CSV file directly. This has 2 consequences:
- most users would prefer this because it avoids the step of having to
unzip the downloaded zip file.
- as there is no manifest, there is no way to show the query criteria
that were used to create the download.

This uses the package `smart_open` which unfortunately uses `boto` and
not `boto3`. That is not too bad, as `boto` and `boto3` can co-exist.