Fix tests with references to the unittest.mock module
This changeset addresses an issue that was uncovered shortly after the updates to the development requirements.  I had not cleaned out my virtual environment entirely and therefore still had the existing backport mock library in it, which prevented me from seeing the import errors with the change to using the internal mock library.  The library needs to now reference unittest.mock instead.

h/t to @vrajmohan for catching this, thank you!  It is also still unclear why the Travis builds have not caught this issue either.