Remove nightly retry processing of itemized schedule A and B data
This changeset removes the nightly retry processing we had initially setup for the itemized schedule A and B data.  We have found this to be more cumbersome to get working and maintain than it is worth and it has resulted in several nightly updates failing due to deadlocked processes and record conflicts.

We will attempt to improve our processing and updating of itemized schedule A and B data through our continued refactoring that we had started previously to make the partitioning code more efficient and easier to maintain.