Add support for multi-column GIN indexes
This changeset modifies our GIN indexes in the itemized schedule A and B data to be multi-column, similar to the other multi-column indexes we have setup.  The intent is to provide a boost in performance when searching/filtering on text/ts_vector fields across multiple cycles.

This requires the support of the btree_gin PostgreSQL extension, which is installed in Amazon RDS but not enabled by default.  It can be enabled by running the following:

CREATE EXTENSION btree_gin;

This will also require repartitioning to rebuild the indexes.