Tweak Celery config to flip ack_late setting
This changeset flips the ack_late (task_acks_late in the config code) setting to be False again due to some strange behavior we are seeing in the dev environment.  Tasks occasionally get run twice and at the moment this appears to be the only culprit.  The documentation in Celery clearly states that if a worker crashes while processing the task will be run again, but we are not seeing a crash when a task has run more than once.

While this may help the issue, we should continue investigating to see if tasks are being redelivered to the workers as that can also cause this to happen.  It seems that the celery flower monitoring tool is a potential culprit in that case but again, it is unclear in our case.

More information can be found in this user group thread:  https://groups.google.com/forum/#!msg/celery-users/W0Qf09ahjas/FgVWqrefkAcJ