Remove nightly refresh of schedule A and B in prod
THis changeset is intended to alleviate an immediate problem we are facing in prod with record inserts.  This work was already completed in the development branch with #2589 and tested in that environment and has had the desired and positive impact we were looking for.  The situation in prod is becoming worse and we cannot wait for the next release for this to be fixed, hence this hotfix.