Add update_functions to nightly refresh
This changeset adds a call to `update_functions` to the nightly process to ensure that the function definitions get updated properly and are not missed.  We have run into a couple of issues in the past with these being out of sync and this is a good way to prevent that from happening again.