Add support for separate partition index management
This changeset adds support to manage the indexes separately in the itemized schedule A and B partitions.  It does this by leveraging the existing code to setup the tables and creates the indexes only if they do not already exist.  After that is finished, any old indexes are then dropped so performance on the tables is not impacted.  The management is exposed as a new management command, `index_itemized`.