Add support for CircleCI service integration
This changeset begins the support for migrating our continuous integration over from Travis CI to CircleCI.  We need to move off of Travis CI per GSA requirements as it is no longer approve, and CircleCI is.  Once this fully working in CircleCI (meaning we can also deploy from CircleCI), we will shut off and remove the Travis CI integration.