Exclude surrogate key from query
Use an explicit query by "known" fields as opposed to a refresh() - the
refresh uses the surrogate key for the query, which may results in rows
not being found as the underlying materialized view is refreshed.

Conjecture: The surrogate `idx` fields is populated using the row_number() window
function. As the populating query does not have an order by, the results
are not repeatable and the row_number may change.

This issue has caused intermittent failures in some environments.