Match virtualenv name in CircleCI config to our ignore files
This changeset updates the CircleCI config to use the same name for the Python virtual environment it creates for testing and deployment as we have in the .gitignore and .cfignore files.  When the name does not match, the virtual environment directory and its contents are pushed up to the cloud.gov environment, which we do not want as it is extra cruft and could potentially cause environment conflicts/issues.