Improve logging setup and update New Relic and requests
This changeset modifies the API logging configuration to ensure that as much info is possible is coming through into our logs to help provide visibility into what is happening at any given moment.  This includes a change to the Flask application configuration as well as the gunicorn startup command.

Furthermore, it also updates the newrelic and requests packages to ensure that we are using the latest versions and accounting for recent bugfixes and enhancements.