Add swagger-tools installation back to README.md
This changeset adds the line to install `swagger-tools` back into the README.md file after we had inadvertently taken it out.  These tools are required in order to run one of the automated tests we have that tests to ensure the API documention is generated correctly.  I have included an extra note in the README about this requirement so folks know why it is there and what purpose it serves.