Fixed refresh materialized management method
This changeset fixes an issue that we uncovered with the latest production deployment, where the nightly updates had partially failed due to materialized views not being refreshed properly.  The cause was because the `db.session` object was being used instead of the `db.engine` object, which in turn made the queries go through the custom router we have setup instead of connecting directly to the master database.